pyproject.toml
[project]
name = "strands-orchestrator"
version = "0.1.0"
description = "Protocol-based DI framework for Strands Agents SDK — agent pooling, MCP management, mode filtering, and extensible hooks"
readme = "README.md"
license = "MIT"
authors = [
    { name = "jtrueblood", email = "jon@mlvx.io" }
]
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "anthropic>=0.86.0",
    "openai>=2.29.0",
]

[project.optional-dependencies]
mongodb = ["motor>=3.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
